+++ src/crc_common.h
@@ -5,9 +5,9 @@
 #define CRC16_POLYNOMIAL 0x1021
 #define CRC32_POLYNOMIAL 0x04C11DB7
 
-#define CRC8_MSB  (1 << 7)
-#define CRC16_MSB (1 << 15)
-#define CRC32_MSB (1 << 31)
+#define CRC8_MSB  ((uint8_t)1 << 7)
+#define CRC16_MSB ((uint16_t)1 << 15)
+#define CRC32_MSB ((uint32_t)1 << 31)
 
 #define CRC16_2_MSB 8
 #define CRC32_2_MSB 24
